requirements.txt
Django>=4.2.0
google-generativeai>=0.3.0
Pillow>=9.0.0
django-widget-tweaks>=1.4.12